# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LumoNAS API
  description: |
    LumoNAS management daemon API. Provides endpoints for system monitoring,
    storage management, Docker orchestration, networking, backups, and administration.
  version: 0.1.0
  contact:
    name: LumoNAS
    url: https://github.com/lumonas/lumonas
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: Local development
...
paths:
...
  /docker/stacks:
    get:
      tags: [Docker]
      summary: List Docker stacks
      operationId: listDockerStacks
      responses:
        '200':
          description: Stack list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DockerStack'
    post:
      tags: [Docker]
      summary: Create a Docker stack
      operationId: createDockerStack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                catalogId:
                  type: string
                values:
                  type: object
                storage:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Stack created
...
  /alerts/{id}/ack:
    patch:
      tags: [Monitoring]
      summary: Acknowledge alert
      operationId: acknowledge-alert
      responses:
        '200':
          description: OK
        '202':
          description: Accepted (async job queued)
        '4XX':
          description: Client error
...
components:
  schemas:
...
    HealthState:
      type: string
      enum: [healthy, attention, warning, critical, offline]
...
    DockerStack:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        status:
          $ref: '#/components/schemas/HealthState'
        state:
          type: string
        images:
          type: array
          items:
            type: string
        composeYaml:
          type: string
        ports:
          type: array
          items:
            type: object
        risks:
          type: array
          items:
            type: string
        recovery:
          $ref: '#/components/schemas/RecoveryContract'
        recoveryCoverage:
          type: number
...
    RecoveryContract:
      type: object
      properties:
        strategy:
          type: string
          enum: [stop-backup, snapshot, custom, none]
        appdataPaths:
          type: array
          items:
            type: string
        preBackupHook:
          $ref: '#/components/schemas/RecoveryHook'
        postBackupHook:
          $ref: '#/components/schemas/RecoveryHook'
        restoreHook:
          $ref: '#/components/schemas/RecoveryHook'
        dbDump:
          $ref: '#/components/schemas/RecoveryHook'
        dbRestore:
          $ref: '#/components/schemas/RecoveryHook'
        stopServices:
          type: array
          items:
            type: string
        startServices:
          type: array
          items:
            type: string
    RecoveryHook:
      type: object
      properties:
        container:
          type: string
        command:
          type: string
        timeout:
          type: integer
...